+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise.js
@@ -1,5 +1,7 @@
 {
     load: 'Loading data ...',
+    reload: 'Loading data ...',
+    update: 'Updating data ...',
     submit: 'Submitting data ...',
     save: 'Updating data ...',
     destroy: 'Updating data ...'