+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_bs.js
@@ -1,5 +1,7 @@
 {
     load: 'Učitavam podatke ...',
+    reload: 'Učitavam podatke ...',
+    update: 'Ažuriranje podataka ...',
     submit: 'Dostavljanje podataka ...',
     save: 'Ažuriranje podataka ...',
     destroy: 'Ažuriranje podataka ...'