+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_cs.js
@@ -1,5 +1,7 @@
 {
     load: 'Načítání dat ...',
+    reload: 'Načítání dat ...',
+    update: 'Aktualizace dat ...',
     submit: 'Odeslání dat ...',
     save: 'Aktualizace dat ...',
     destroy: 'Aktualizace dat ...'