+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_es.js
@@ -1,5 +1,7 @@
 {
     load: 'Cargando datos ...',
+    reload: 'Cargando datos ...',
+    update: 'Actualización de datos ...',
     submit: 'Presentación de los datos ...',
     save: 'Actualización de datos ...',
     destroy: 'Actualización de datos ...'