+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_fi.js
@@ -1,5 +1,7 @@
 {
     load: 'Ladataan tietoja ...',
+    reload: 'Ladataan tietoja ...',
+    update: 'Päivittää tietoja ...',
     submit: 'Toimittaa tiedot ...',
     save: 'Päivittää tietoja ...',
     destroy: 'Päivittää tietoja ...'