+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_fr.js
@@ -1,5 +1,7 @@
 {
     load: 'Chargement des données ...',
+    reload: 'Chargement des données ...',
+    update: 'Mise à jour des données ...',
     submit: 'Envoi de données ...',
     save: 'Mise à jour des données ...',
     destroy: 'Mise à jour des données ...'