+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_hr.js
@@ -1,5 +1,7 @@
 {
     load: 'Učitavanje podataka ...',
+    reload: 'Učitavanje podataka ...',
+    update: 'Ažuriranje podataka ...',
     submit: 'Slanje podataka ...',
     save: 'Ažuriranje podataka ...',
     destroy: 'Ažuriranje podataka ...'