+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_hu.js
@@ -1,5 +1,7 @@
 {
     load: 'Adatok betöltése ...',
+    reload: 'Adatok betöltése ...',
+    update: 'Adatok frissítése ...',
     submit: 'Adatok küldése ...',
     save: 'Adatok frissítése ...',
     destroy: 'Adatok frissítése ...'