+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_it.js
@@ -1,5 +1,7 @@
 {
     load: 'Caricamento dei dati ...',
+    reload: 'Caricamento dei dati ...',
+    update: 'Aggiornamento dei dati ...',
     submit: 'Invio dei dati ...',
     save: 'Aggiornamento dei dati ...',
     destroy: 'Aggiornamento dei dati ...'