+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_nl.js
@@ -1,6 +1,8 @@
 {
     load: 'Laden gegevens ...',
+    reload: 'Laden gegevens ...',
+    update: 'Updatingdata',
     submit: 'Versturen gegevens ...',
-    save: 'Bijwerken gegevens ...',
+    save: 'Bewaren gegevens ...',
     destroy: 'Bijwerken gegevens ...'
 }