+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_pl.js
@@ -1,5 +1,7 @@
 {
     load: 'Ładowanie danych ...',
+    reload: 'Ładowanie danych ...',
+    update: 'Aktualizacja danych ...',
     submit: 'Przesyłanie danych ...',
     save: 'Aktualizacja danych ...',
     destroy: 'Aktualizacja danych ...'